push

+++ printf/ft_puthexa.c
@@ -6,14 +6,14 @@
 /*   By: azrig <azrig@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/04 14:56:08 by azrig             #+#    #+#             */
-/*   Updated: 2025/01/05 22:44:31 by azrig            ###   ########.fr       */
+/*   Updated: 2025/01/06 00:23:53 by azrig            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include <stdio.h>
 
-int	ft_puthexa(unsigned int n, char format)
+int	ft_puthexa(unsigned long n, char format)
 {
 	char	*alpha;
 	int		count;
